requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
xgboost
lightgbm
imbalanced-learn
joblib
matplotlib
torch
torch-geometric
rdkit-pypi
shap
openpyxl
